+++ src/main/java/noppes/mpm/client/EmoteHandler.java
@@ -1,3 +1,4 @@
+package noppes.mpm.client;
 /**
  * This class was created by <Vazkii>. It's distributed as
  * part of the Quark Mod. Get the Source Code in github:
@@ -8,7 +9,6 @@
  *
  * File Created @ [26/03/2016, 21:37:17 (GMT)]
  */
-package main.java.noppes.mpm.client;
 
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.entity.AbstractClientPlayer;
@@ -28,20 +28,21 @@ import net.minecraft.world.World;
 import net.minecraftforge.fml.common.ObfuscationReflectionHelper;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
-import vazkii.quark.base.client.ModKeybinds;
-import vazkii.quark.base.lib.LibObfuscation;
-import vazkii.quark.base.network.message.MessageRequestEmote;
-import vazkii.quark.vanity.feature.EmoteSystem;
-import vazkii.aurelienribon.tweenengine.*;
+
+import aurelienribon.tweenengine.*;
+import noppes.mpm.client.ModelAccessor;
 
 import java.io.IOException;
 import java.io.File;
 import java.io.FileReader;
 import java.io.BufferedReader;
+import java.util.Map;
 import java.util.HashMap;
 import java.util.LinkedHashMap;
-import java.util.Map;
 import java.util.ArrayList;
+import java.lang.reflect.Field;
+import java.lang.reflect.Modifier;
+import java.util.*;
 
 public final class EmoteHandler {
 	public static enum ASTType {
@@ -60,6 +61,7 @@ public final class EmoteHandler {
 	public static class AST {
 		public ASTType type;
 		public String str;
+		public int i;
 		public Float numeral0;
 		public Float numeral1;
 		public AST child0;//NOTE: we can optimize this by compiling to bytecode instead of an ast (java makes this very hard to do efficiently)
@@ -72,14 +74,55 @@ public final class EmoteHandler {
 
 		public float animSpeed;
 		public ArrayList<Integer> usedParts;
-		public AST ast;
+		public AST root;
 	}
 
 	public static final Map<String, EmoteDescriptor> emoteMap = new LinkedHashMap<>();
+	private static final Map<String, TweenEquation> equationMap = new HashMap<>();
+	private static final Map<String, Integer> tweenableMap = new HashMap<>();
+	private static final Map<String, Integer> partMap = new HashMap<>();
 	// private static final Map<String, EmoteBase> playerEmotes = new HashMap<>();
 	// private static final Map<String, EmoteBase> playerWalks = new HashMap<>();
 
 	private static int count;
+
+	static {
+		Class<?> clazz = ModelAccessor.class;
+		Field[] fields = clazz.getDeclaredFields();
+		for(Field f : fields) {
+			if(f.getType() != int.class)
+				continue;
+
+			int modifiers = f.getModifiers();
+			if(Modifier.isPublic(modifiers) && Modifier.isStatic(modifiers) && Modifier.isFinal(modifiers)) {
+				try {
+					int val = f.getInt(null);
+					String name = f.getName().toLowerCase();
+					if(name.matches("^.+?_[xyz]$"))
+						tweenableMap.put(name, val);
+					else
+						partMap.put(name, val);
+				} catch(Exception e) {
+					e.printStackTrace();
+				}
+			}
+		}
+
+		clazz = TweenEquations.class;
+		fields = clazz.getDeclaredFields();
+		for(Field f : fields) {
+			String name = f.getName().replaceAll("[A-Z]", "_$0").substring(5).toLowerCase();
+			try {
+				TweenEquation eq = (TweenEquation) f.get(null);
+				equationMap.put(name, eq);
+				if(name.equals("none")) {
+					equationMap.put("linear", eq);
+				}
+			} catch(Exception e) {
+				e.printStackTrace();
+			}
+		}
+	}
 
 	public static boolean addEmote(String name) {
 		EmoteDescriptor desc = new EmoteDescriptor();
@@ -114,7 +157,7 @@ public final class EmoteHandler {
 
 		int[] token_i = {0};
 		String[] error_msg = {null};
-		desc.ast = parse_top_level(emoteTokens, token_i, error_msg, desc);
+		desc.root = parse_top_level(emoteTokens, token_i, error_msg, desc);
 		if(error_msg[0] != null) {
 			//TODO: add error logging
 			return false;
@@ -144,13 +187,13 @@ public final class EmoteHandler {
 
 	public static AST parse_top_level(ArrayList<String> tokens, int[] token_i, String[] error_msg, EmoteDescriptor desc) {
 		if(eat_token(tokens, token_i, "use")) {
-			String part = tokens.get(token_i[0]++);
-			if(part.equals("EOF\n")) {
-				error_msg[0] = msg_unexpected("EOF", "string");
+			String token = tokens.get(token_i[0]++);
+			if(partMap.containsKey(token)) {
+				desc.usedParts.add(partMap.get(token));
+			} else {
+				error_msg[0] = msg_unexpected(token, "part name");
 				return null;
 			}
-			//TODO: add to parts list
-			// desc.usedParts.add(part);
 
 			eat_token(tokens, token_i, "\n");
 			return parse_top_level(tokens, token_i, error_msg, desc);
@@ -160,7 +203,7 @@ public final class EmoteHandler {
 			try {
 				speed = Float.parseFloat(token);
 			} catch(NumberFormatException e) {
-				error_msg[0] = msg_unexpected(tokens.get(token_i[0] - 1), "number");
+				error_msg[0] = msg_unexpected(token, "number");
 				return null;
 			}
 			desc.animSpeed = speed;
@@ -171,6 +214,7 @@ public final class EmoteHandler {
 			AST ast = parse_animation(tokens, token_i, error_msg);
 			if(ast == null) {
 				error_msg[0] = msg_unexpected(tokens.get(token_i[0]), "keyword: 'use', 'unit' or 'animation'");
+				return null;
 			}
 			return ast;
 		}
@@ -184,7 +228,7 @@ public final class EmoteHandler {
 			} else if(eat_token(tokens, token_i, "parallel")) {
 				ast.type = ASTType.PARALLEL;
 			} else {
-				error_msg[0] = msg_unexpected(tokens.get(token_i[0] - 1), "animation type ('sequence' or 'parallel')");
+				error_msg[0] = msg_unexpected(tokens.get(token_i[0]), "animation type ('sequence' or 'parallel')");
 				return null;
 			}
 
@@ -209,18 +253,19 @@ public final class EmoteHandler {
 			AST ast = new AST();
 			ast.type = ASTType.MOVE;
 
-			ast.str = tokens.get(token_i[0]++);
-			//TODO: check tweenable parts
-			if(ast.str.equals("EOF\n")) {
-				error_msg[0] = msg_unexpected("EOF", "part name");
+			String token = tokens.get(token_i[0]++);
+			if(tweenableMap.containsKey(token)) {
+				ast.i = tweenableMap.get(token);
+			} else {
+				error_msg[0] = msg_unexpected(token, "part name");
 				return null;
 			}
 
-			String token = tokens.get(token_i[0]++);
+			token = tokens.get(token_i[0]++);
 			try {
 				ast.numeral0 = Float.parseFloat(token);
 			} catch(NumberFormatException e) {
-				error_msg[0] = msg_unexpected(tokens.get(token_i[0] - 1), "time duration");
+				error_msg[0] = msg_unexpected(token, "time duration");
 				return null;
 			}
 
@@ -228,7 +273,7 @@ public final class EmoteHandler {
 			try {
 				ast.numeral1 = Float.parseFloat(token);
 			} catch(NumberFormatException e) {
-				error_msg[0] = msg_unexpected(tokens.get(token_i[0] - 1), "target rotation");
+				error_msg[0] = msg_unexpected(token, "target rotation");
 				return null;
 			}
 
@@ -241,7 +286,7 @@ public final class EmoteHandler {
 					try {
 						cmd_ast.numeral0 = Float.parseFloat(token);
 					} catch(NumberFormatException e) {
-						error_msg[0] = msg_unexpected(tokens.get(token_i[0] - 1), "delay duration");
+						error_msg[0] = msg_unexpected(token, "delay duration");
 						return null;
 					}
 					parent.child1 = cmd_ast;
@@ -255,16 +300,16 @@ public final class EmoteHandler {
 					}
 					token = tokens.get(token_i[0]++);
 					try {
-						cmd_ast.numeral0 = 1.0f*Integer.parseInt(token);
+						cmd_ast.i = Integer.parseInt(token);
 					} catch(NumberFormatException e) {
-						error_msg[0] = msg_unexpected(tokens.get(token_i[0] - 1), "number of times to repeat");
+						error_msg[0] = msg_unexpected(token, "number of times to repeat");
 						return null;
 					}
 					token = tokens.get(token_i[0]++);
 					try {
-						cmd_ast.numeral1 = Float.parseFloat(token);
+						cmd_ast.numeral0 = Float.parseFloat(token);
 					} catch(NumberFormatException e) {
-						error_msg[0] = msg_unexpected(tokens.get(token_i[0] - 1), "delay duration");
+						error_msg[0] = msg_unexpected(token, "delay duration");
 						return null;
 					}
 					parent.child1 = cmd_ast;
@@ -273,11 +318,10 @@ public final class EmoteHandler {
 					AST cmd_ast = new AST();
 					cmd_ast.type = ASTType.MOVE_EASE;
 					cmd_ast.str = tokens.get(token_i[0]++);
-					//TODO: add equation checking
-					// if(!equations.containsKey(cmd_ast.str)) {
-					// 	error_msg[0] = msg_unexpected(tokens.get(token_i[0] - 1), "valid easing type");
-					// 	return null;
-					// }
+					if(!equationMap.containsKey(cmd_ast.str)) {
+						error_msg[0] = msg_unexpected(cmd_ast.str, "valid easing type");
+						return null;
+					}
 
 					parent.child1 = cmd_ast;
 					parent = cmd_ast;
@@ -295,13 +339,17 @@ public final class EmoteHandler {
 			AST ast = new AST();
 			ast.type = ASTType.RESET;
 
-			ast.str = tokens.get(token_i[0]++);
-			//TODO: check tweenable parts or if str == all
-			if(ast.str.equals("EOF\n")) {
-				error_msg[0] = msg_unexpected("EOF", "part name");
+			String token = tokens.get(token_i[0]++);
+			if(token.equals("all")) {
+				ast.i = 0;
+			} else if(tweenableMap.containsKey(token)) {
+				ast.i = tweenableMap.get(token);
+			} else {
+				error_msg[0] = msg_unexpected(token, "part name");
 				return null;
 			}
-			String token = tokens.get(token_i[0]++);
+
+			token = tokens.get(token_i[0]++);
 			if(token.equals("all")) {
 				ast.numeral0 = 1.0f;
 			} else if(token.equals("rotation")) {
@@ -309,7 +357,15 @@ public final class EmoteHandler {
 			} else if(token.equals("offset")) {
 				ast.numeral0 = 3.0f;
 			} else {
-				error_msg[0] = msg_unexpected(tokens.get(token_i[0] - 1), "keyword 'all', 'rotation' or 'offset'");
+				error_msg[0] = msg_unexpected(token, "keyword 'all', 'rotation' or 'offset'");
+				return null;
+			}
+
+			token = tokens.get(token_i[0]++);
+			try {
+				ast.numeral1 = Float.parseFloat(token);
+			} catch(NumberFormatException e) {
+				error_msg[0] = msg_unexpected(token, "reset duration");
 				return null;
 			}
 
@@ -323,7 +379,7 @@ public final class EmoteHandler {
 			try {
 				ast.numeral0 = Float.parseFloat(token);
 			} catch(NumberFormatException e) {
-				error_msg[0] = msg_unexpected(tokens.get(token_i[0] - 1), "pause duration");
+				error_msg[0] = msg_unexpected(token, "pause duration");
 				return null;
 			}
 
@@ -333,22 +389,22 @@ public final class EmoteHandler {
 		} else if(eat_token(tokens, token_i, "yoyo") || eat_token(tokens, token_i, "repeat")) {
 			AST ast = new AST();
 			if(tokens.get(token_i[0] - 1).equals("yoyo")) {
-				ast.type = ASTType.MOVE_YOYO;
+				ast.type = ASTType.YOYO;
 			} else {
-				ast.type = ASTType.MOVE_REPEAT;
+				ast.type = ASTType.REPEAT;
 			}
 			String token = tokens.get(token_i[0]++);
 			try {
-				ast.numeral0 = 1.0f*Integer.parseInt(token);
+				ast.i = Integer.parseInt(token);
 			} catch(NumberFormatException e) {
-				error_msg[0] = msg_unexpected(tokens.get(token_i[0] - 1), "number of times to repeat");
+				error_msg[0] = msg_unexpected(token, "number of times to repeat");
 				return null;
 			}
 			token = tokens.get(token_i[0]++);
 			try {
-				ast.numeral1 = Float.parseFloat(token);
+				ast.numeral0 = Float.parseFloat(token);
 			} catch(NumberFormatException e) {
-				error_msg[0] = msg_unexpected(tokens.get(token_i[0] - 1), "delay duration");
+				error_msg[0] = msg_unexpected(token, "delay duration");
 				return null;
 			}
 
@@ -357,6 +413,113 @@ public final class EmoteHandler {
 			return ast;
 		} else {
 			return parse_animation(tokens, token_i, error_msg);
+		}
+	}
+
+	public static Timeline creatTimeline(EmoteDescriptor desc, ModelBiped model) {
+		if(desc.root.type == ASTType.SEQUENCE) {
+			Timeline timeline = Timeline.createSequence();
+			buildTimeline(timeline, desc.root.child0, desc, model);
+			return timeline;
+		} else if(desc.root.type == ASTType.PARALLEL) {
+			Timeline timeline = Timeline.createParallel();
+			buildTimeline(timeline, desc.root.child0, desc, model);
+			return timeline;
+		} else {
+			//TODO: assertion logging
+			Timeline timeline = Timeline.createParallel();
+			return timeline;
+		}
+	}
+	public static void buildTimeline(Timeline timeline, AST ast, EmoteDescriptor desc, ModelBiped model) {
+		if(ast.type == ASTType.SEQUENCE) {
+			Timeline newTimeline = Timeline.createSequence();
+			buildTimeline(newTimeline, ast.child0, desc, model);
+			timeline.push(newTimeline);
+		} else if(ast.type == ASTType.PARALLEL) {
+			Timeline newTimeline = Timeline.createParallel();
+			buildTimeline(newTimeline, ast.child0, desc, model);
+			timeline.push(newTimeline);
+		} else if(ast.type == ASTType.MOVE) {
+			int part = tweenableMap.get(ast.str);
+			Tween tween = Tween.to(model, part, ast.numeral0*desc.animSpeed).target(ast.numeral1);
+			buildTween(tween, ast.child1, desc.animSpeed);
+			timeline.push(tween);
+		} else if(ast.type == ASTType.RESET) {
+			int part = ast.i;
+			float time = ast.numeral1*desc.animSpeed;
+			boolean all = ast.numeral0 == 1.0f;
+			boolean rot = all || ast.numeral0 == 2.0f;
+			boolean off = all || ast.numeral0 == 3.0f;
+
+			Timeline parallel = Timeline.createParallel();
+			int lower = (part == 0) ? 0 : part + (rot ? 0 : 3);
+			int upper = (part == 0) ? ModelAccessor.STATE_COUNT : part + (off ? ModelAccessor.STATE_COUNT : 3);
+
+			for(int i = lower; i < upper; i++) {
+				int piece = (i / ModelAccessor.MODEL_PROPS) * ModelAccessor.MODEL_PROPS;
+				if(desc.usedParts.contains(piece)) parallel.push(Tween.to(model, i, time));
+			}
+
+			timeline.push(parallel);
+		} else if(ast.type == ASTType.PAUSE) {
+			timeline.pushPause(ast.numeral0*desc.animSpeed);
+		} else if(ast.type == ASTType.YOYO) {
+			timeline.repeatYoyo(ast.i, ast.numeral0*desc.animSpeed);
+		} else if(ast.type == ASTType.REPEAT) {
+			timeline.repeat(ast.i, ast.numeral0*desc.animSpeed);
+		} else {
+			//TODO: assertion logging
+		}
+	}
+
+	public static void buildTween(Tween tween, AST ast, Float speed) {
+		if(ast == null) return;
+
+		if(ast.type == ASTType.MOVE_DELAY) {
+			tween.delay(ast.numeral0*speed);
+			buildTween(tween, ast.child1, speed);
+		} else if(ast.type == ASTType.MOVE_YOYO) {
+			tween.repeatYoyo(ast.i, ast.numeral0*speed);
+			buildTween(tween, ast.child1, speed);
+		} else if(ast.type == ASTType.MOVE_REPEAT) {
+			tween.repeat(ast.i, ast.numeral0*speed);
+			buildTween(tween, ast.child1, speed);
+		} else if(ast.type == ASTType.MOVE_EASE) {
+			tween.ease(equationMap.get(ast.str));
+			buildTween(tween, ast.child1, speed);
+		} else {
+			//TODO: assertion logging
+		}
+	}
+
+
+	public static boolean attemptEmote(EntityPlayer player, String name) {
+		// String name = player.getName();
+		EmoteDescriptor desc = emoteMap.get(name);
+		if(desc == null) return false;
+
+		ModelBiped model;
+		{//get player model
+			Minecraft mc = Minecraft.getMinecraft();
+			RenderManager manager = mc.getRenderManager();
+			RenderPlayer render = manager.getSkinMap().get(((AbstractClientPlayer) player).getSkinType());
+			model = render.getMainModel();
+		}
+
+		if(model != null) {
+			// resetPlayer(player);
+
+			Timeline timeline = creatTimeline(desc, model);
+
+			timeline.start(player);
+			// lastMs = System.currentTimeMillis();
+			// playerEmotes.put(name, emote);
+			// if (desc.getWalkstyle() == true)
+			// 	playerWalks.put(name, emote);
+			return true;
+		} else {
+			return false;
 		}
 	}
 }
